Municipalidad/requirements.txt

django-crispy-forms
crispy-bootstrap4
